+++ src/Firewall/Rules/CountryRule.cs
@@ -1,10 +1,13 @@
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using MaxMind.Db;
 using MaxMind.GeoIP2;
+using MaxMind.GeoIP2.Exceptions;
 using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Logging;
 
 namespace Firewall
 {
@@ -25,7 +28,10 @@ namespace Firewall
             IList<CountryCode> allowedCountries,
             string geoIP2FileName = null)
         {
-            _nextRule = nextRule;
+            _nextRule = nextRule ?? throw new ArgumentNullException(nameof(nextRule));
+            if (allowedCountries == null)
+                throw new ArgumentNullException(nameof(allowedCountries));
+
             _allowedCountries =
                 allowedCountries
                     .Select(isoCode => isoCode.ToString())
@@ -47,11 +53,31 @@ namespace Firewall
         /// </summary>
         public bool IsAllowed(HttpContext context)
         {
-            var remoteIpAddress = context.Connection.RemoteIpAddress;
-            var result = _databaseReader.Country(remoteIpAddress);
-            var countryCode = result.Country.IsoCode;
-            return _allowedCountries.Contains(countryCode) || _nextRule.IsAllowed(context);
+            try
+            {
+                var remoteIpAddress = context.Connection.RemoteIpAddress;
+                var result = _databaseReader.Country(remoteIpAddress);
+                var countryCode = result.Country.IsoCode;
 
+                var isAllowed = _allowedCountries.Contains(countryCode);
+
+                context.LogDebug(
+                    nameof(CountryRule),
+                    isAllowed,
+                    "it originated in '{country}'",
+                    result.Country.Name);
+
+                return isAllowed || _nextRule.IsAllowed(context);
+            }
+            catch (AddressNotFoundException)
+            {
+                context.LogDebug(
+                    nameof(CountryRule),
+                    false,
+                    "it couldn't be verified against the current GeoIP2 database.");
+
+                return _nextRule.IsAllowed(context);
+            }
         }
     }
 }